🏗️ Déplacement du chargement dans le Provider
déplacement du chargement des paramètre de ParamaScren à ConfigurationProvider afin que le chargement se fait des le chargement de l'application
